Refactor TerminalImportSession out of beets.ui.commands
