+++ beets/ui/terminal_import_session.py
@@ -1,0 +1,715 @@
+"""An import session that runs in a terminal.
+"""
+
+from __future__ import (division, absolute_import, print_function,
+                        unicode_literals)
+
+import os
+
+from beets import ui
+from beets.ui import print_, input_
+from beets import autotag
+from beets.autotag import Recommendation
+from beets.autotag import hooks
+from beets import importer
+from beets.util import displayable_path
+from beets import config
+from beets import logging
+
+VARIOUS_ARTISTS = u'Various Artists'
+
+# Global logger.
+log = logging.getLogger('beets')
+
+
+# Importer utilities and support.
+
+def disambig_string(info):
+    """Generate a string for an AlbumInfo or TrackInfo object that
+    provides context that helps disambiguate similar-looking albums and
+    tracks.
+    """
+    disambig = []
+    if info.data_source and info.data_source != 'MusicBrainz':
+        disambig.append(info.data_source)
+
+    if isinstance(info, hooks.AlbumInfo):
+        if info.media:
+            if info.mediums > 1:
+                disambig.append(u'{0}x{1}'.format(
+                    info.mediums, info.media
+                ))
+            else:
+                disambig.append(info.media)
+        if info.year:
+            disambig.append(unicode(info.year))
+        if info.country:
+            disambig.append(info.country)
+        if info.label:
+            disambig.append(info.label)
+        if info.albumdisambig:
+            disambig.append(info.albumdisambig)
+
+    if disambig:
+        return u', '.join(disambig)
+
+
+def dist_string(dist):
+    """Formats a distance (a float) as a colorized similarity percentage
+    string.
+    """
+    out = '%.1f%%' % ((1 - dist) * 100)
+    if dist <= config['match']['strong_rec_thresh'].as_number():
+        out = ui.colorize('text_success', out)
+    elif dist <= config['match']['medium_rec_thresh'].as_number():
+        out = ui.colorize('text_warning', out)
+    else:
+        out = ui.colorize('text_error', out)
+    return out
+
+
+def penalty_string(distance, limit=None):
+    """Returns a colorized string that indicates all the penalties
+    applied to a distance object.
+    """
+    penalties = []
+    for key in distance.keys():
+        key = key.replace('album_', '')
+        key = key.replace('track_', '')
+        key = key.replace('_', ' ')
+        penalties.append(key)
+    if penalties:
+        if limit and len(penalties) > limit:
+            penalties = penalties[:limit] + ['...']
+        return ui.colorize('text_warning', '(%s)' % ', '.join(penalties))
+
+
+def show_change(cur_artist, cur_album, match):
+    """Print out a representation of the changes that will be made if an
+    album's tags are changed according to `match`, which must be an AlbumMatch
+    object.
+    """
+    def show_album(artist, album):
+        if artist:
+            album_description = u'    %s - %s' % (artist, album)
+        elif album:
+            album_description = u'    %s' % album
+        else:
+            album_description = u'    (unknown album)'
+        print_(album_description)
+
+    def format_index(track_info):
+        """Return a string representing the track index of the given
+        TrackInfo or Item object.
+        """
+        if isinstance(track_info, hooks.TrackInfo):
+            index = track_info.index
+            medium_index = track_info.medium_index
+            medium = track_info.medium
+            mediums = match.info.mediums
+        else:
+            index = medium_index = track_info.track
+            medium = track_info.disc
+            mediums = track_info.disctotal
+        if config['per_disc_numbering']:
+            if mediums > 1:
+                return u'{0}-{1}'.format(medium, medium_index)
+            else:
+                return unicode(medium_index)
+        else:
+            return unicode(index)
+
+    # Identify the album in question.
+    if cur_artist != match.info.artist or \
+            (cur_album != match.info.album and
+             match.info.album != VARIOUS_ARTISTS):
+        artist_l, artist_r = cur_artist or '', match.info.artist
+        album_l,  album_r = cur_album or '', match.info.album
+        if artist_r == VARIOUS_ARTISTS:
+            # Hide artists for VA releases.
+            artist_l, artist_r = u'', u''
+
+        artist_l, artist_r = ui.colordiff(artist_l, artist_r)
+        album_l, album_r = ui.colordiff(album_l, album_r)
+
+        print_("Correcting tags from:")
+        show_album(artist_l, album_l)
+        print_("To:")
+        show_album(artist_r, album_r)
+    else:
+        print_(u"Tagging:\n    {0.artist} - {0.album}".format(match.info))
+
+    # Data URL.
+    if match.info.data_url:
+        print_('URL:\n    %s' % match.info.data_url)
+
+    # Info line.
+    info = []
+    # Similarity.
+    info.append('(Similarity: %s)' % dist_string(match.distance))
+    # Penalties.
+    penalties = penalty_string(match.distance)
+    if penalties:
+        info.append(penalties)
+    # Disambiguation.
+    disambig = disambig_string(match.info)
+    if disambig:
+        info.append(ui.colorize('text_highlight_minor', '(%s)' % disambig))
+    print_(' '.join(info))
+
+    # Tracks.
+    pairs = match.mapping.items()
+    pairs.sort(key=lambda arg: arg[1].index)
+
+    # Build up LHS and RHS for track difference display. The `lines` list
+    # contains ``(lhs, rhs, width)`` tuples where `width` is the length (in
+    # characters) of the uncolorized LHS.
+    lines = []
+    medium = disctitle = None
+    for item, track_info in pairs:
+
+        # Medium number and title.
+        if medium != track_info.medium or disctitle != track_info.disctitle:
+            media = match.info.media or 'Media'
+            if match.info.mediums > 1 and track_info.disctitle:
+                lhs = '%s %s: %s' % (media, track_info.medium,
+                                     track_info.disctitle)
+            elif match.info.mediums > 1:
+                lhs = '%s %s' % (media, track_info.medium)
+            elif track_info.disctitle:
+                lhs = '%s: %s' % (media, track_info.disctitle)
+            else:
+                lhs = None
+            if lhs:
+                lines.append((lhs, '', 0))
+            medium, disctitle = track_info.medium, track_info.disctitle
+
+        # Titles.
+        new_title = track_info.title
+        if not item.title.strip():
+            # If there's no title, we use the filename.
+            cur_title = displayable_path(os.path.basename(item.path))
+            lhs, rhs = cur_title, new_title
+        else:
+            cur_title = item.title.strip()
+            lhs, rhs = ui.colordiff(cur_title, new_title)
+        lhs_width = len(cur_title)
+
+        # Track number change.
+        cur_track, new_track = format_index(item), format_index(track_info)
+        if cur_track != new_track:
+            if item.track in (track_info.index, track_info.medium_index):
+                color = 'text_highlight_minor'
+            else:
+                color = 'text_highlight'
+            templ = ui.colorize(color, u' (#{0})')
+            lhs += templ.format(cur_track)
+            rhs += templ.format(new_track)
+            lhs_width += len(cur_track) + 4
+
+        # Length change.
+        if item.length and track_info.length and \
+                abs(item.length - track_info.length) > \
+                config['ui']['length_diff_thresh'].as_number():
+            cur_length = ui.human_seconds_short(item.length)
+            new_length = ui.human_seconds_short(track_info.length)
+            templ = ui.colorize('text_highlight', u' ({0})')
+            lhs += templ.format(cur_length)
+            rhs += templ.format(new_length)
+            lhs_width += len(cur_length) + 3
+
+        # Penalties.
+        penalties = penalty_string(match.distance.tracks[track_info])
+        if penalties:
+            rhs += ' %s' % penalties
+
+        if lhs != rhs:
+            lines.append((' * %s' % lhs, rhs, lhs_width))
+        elif config['import']['detail']:
+            lines.append((' * %s' % lhs, '', lhs_width))
+
+    # Print each track in two columns, or across two lines.
+    col_width = (ui.term_width() - len(''.join([' * ', ' -> ']))) // 2
+    if lines:
+        max_width = max(w for _, _, w in lines)
+        for lhs, rhs, lhs_width in lines:
+            if not rhs:
+                print_(lhs)
+            elif max_width > col_width:
+                print_(u'%s ->\n   %s' % (lhs, rhs))
+            else:
+                pad = max_width - lhs_width
+                print_(u'%s%s -> %s' % (lhs, ' ' * pad, rhs))
+
+    # Missing and unmatched tracks.
+    if match.extra_tracks:
+        print_('Missing tracks ({0}/{1} - {2:.1%}):'.format(
+               len(match.extra_tracks),
+               len(match.info.tracks),
+               len(match.extra_tracks) / len(match.info.tracks)
+               ))
+    for track_info in match.extra_tracks:
+        line = ' ! %s (#%s)' % (track_info.title, format_index(track_info))
+        if track_info.length:
+            line += ' (%s)' % ui.human_seconds_short(track_info.length)
+        print_(ui.colorize('text_warning', line))
+    if match.extra_items:
+        print_('Unmatched tracks ({0}):'.format(len(match.extra_items)))
+    for item in match.extra_items:
+        line = ' ! %s (#%s)' % (item.title, format_index(item))
+        if item.length:
+            line += ' (%s)' % ui.human_seconds_short(item.length)
+        print_(ui.colorize('text_warning', line))
+
+
+def show_item_change(item, match):
+    """Print out the change that would occur by tagging `item` with the
+    metadata from `match`, a TrackMatch object.
+    """
+    cur_artist, new_artist = item.artist, match.info.artist
+    cur_title, new_title = item.title, match.info.title
+
+    if cur_artist != new_artist or cur_title != new_title:
+        cur_artist, new_artist = ui.colordiff(cur_artist, new_artist)
+        cur_title, new_title = ui.colordiff(cur_title, new_title)
+
+        print_("Correcting track tags from:")
+        print_("    %s - %s" % (cur_artist, cur_title))
+        print_("To:")
+        print_("    %s - %s" % (new_artist, new_title))
+
+    else:
+        print_("Tagging track: %s - %s" % (cur_artist, cur_title))
+
+    # Data URL.
+    if match.info.data_url:
+        print_('URL:\n    %s' % match.info.data_url)
+
+    # Info line.
+    info = []
+    # Similarity.
+    info.append('(Similarity: %s)' % dist_string(match.distance))
+    # Penalties.
+    penalties = penalty_string(match.distance)
+    if penalties:
+        info.append(penalties)
+    # Disambiguation.
+    disambig = disambig_string(match.info)
+    if disambig:
+        info.append(ui.colorize('text_highlight_minor', '(%s)' % disambig))
+    print_(' '.join(info))
+
+
+def summarize_items(items, singleton):
+    """Produces a brief summary line describing a set of items. Used for
+    manually resolving duplicates during import.
+
+    `items` is a list of `Item` objects. `singleton` indicates whether
+    this is an album or single-item import (if the latter, them `items`
+    should only have one element).
+    """
+    summary_parts = []
+    if not singleton:
+        summary_parts.append("{0} items".format(len(items)))
+
+    format_counts = {}
+    for item in items:
+        format_counts[item.format] = format_counts.get(item.format, 0) + 1
+    if len(format_counts) == 1:
+        # A single format.
+        summary_parts.append(items[0].format)
+    else:
+        # Enumerate all the formats by decreasing frequencies:
+        for fmt, count in sorted(format_counts.items(),
+                                 key=lambda arg: (-arg[1], arg[0])):
+            summary_parts.append('{0} {1}'.format(fmt, count))
+
+    if items:
+        average_bitrate = sum([item.bitrate for item in items]) / len(items)
+        total_duration = sum([item.length for item in items])
+        total_filesize = sum([item.filesize for item in items])
+        summary_parts.append('{0}kbps'.format(int(average_bitrate / 1000)))
+        summary_parts.append(ui.human_seconds_short(total_duration))
+        summary_parts.append(ui.human_bytes(total_filesize))
+
+    return ', '.join(summary_parts)
+
+
+def _summary_judment(rec):
+    """Determines whether a decision should be made without even asking
+    the user. This occurs in quiet mode and when an action is chosen for
+    NONE recommendations. Return an action or None if the user should be
+    queried. May also print to the console if a summary judgment is
+    made.
+    """
+    if config['import']['quiet']:
+        if rec == Recommendation.strong:
+            return importer.action.APPLY
+        else:
+            action = config['import']['quiet_fallback'].as_choice({
+                'skip': importer.action.SKIP,
+                'asis': importer.action.ASIS,
+            })
+
+    elif rec == Recommendation.none:
+        action = config['import']['none_rec_action'].as_choice({
+            'skip': importer.action.SKIP,
+            'asis': importer.action.ASIS,
+            'ask': None,
+        })
+
+    else:
+        return None
+
+    if action == importer.action.SKIP:
+        print_('Skipping.')
+    elif action == importer.action.ASIS:
+        print_('Importing as-is.')
+    return action
+
+
+def choose_candidate(candidates, singleton, rec, cur_artist=None,
+                     cur_album=None, item=None, itemcount=None):
+    """Given a sorted list of candidates, ask the user for a selection
+    of which candidate to use. Applies to both full albums and
+    singletons  (tracks). Candidates are either AlbumMatch or TrackMatch
+    objects depending on `singleton`. for albums, `cur_artist`,
+    `cur_album`, and `itemcount` must be provided. For singletons,
+    `item` must be provided.
+
+    Returns the result of the choice, which may SKIP, ASIS, TRACKS, or
+    MANUAL or a candidate (an AlbumMatch/TrackMatch object).
+    """
+    # Sanity check.
+    if singleton:
+        assert item is not None
+    else:
+        assert cur_artist is not None
+        assert cur_album is not None
+
+    # Zero candidates.
+    if not candidates:
+        if singleton:
+            print_("No matching recordings found.")
+            opts = ('Use as-is', 'Skip', 'Enter search', 'enter Id',
+                    'aBort')
+        else:
+            print_("No matching release found for {0} tracks."
+                   .format(itemcount))
+            print_('For help, see: '
+                   'http://beets.readthedocs.org/en/latest/faq.html#nomatch')
+            opts = ('Use as-is', 'as Tracks', 'Group albums', 'Skip',
+                    'Enter search', 'enter Id', 'aBort')
+        sel = ui.input_options(opts)
+        if sel == 'u':
+            return importer.action.ASIS
+        elif sel == 't':
+            assert not singleton
+            return importer.action.TRACKS
+        elif sel == 'e':
+            return importer.action.MANUAL
+        elif sel == 's':
+            return importer.action.SKIP
+        elif sel == 'b':
+            raise importer.ImportAbort()
+        elif sel == 'i':
+            return importer.action.MANUAL_ID
+        elif sel == 'g':
+            return importer.action.ALBUMS
+        else:
+            assert False
+
+    # Is the change good enough?
+    bypass_candidates = False
+    if rec != Recommendation.none:
+        match = candidates[0]
+        bypass_candidates = True
+
+    while True:
+        # Display and choose from candidates.
+        require = rec <= Recommendation.low
+
+        if not bypass_candidates:
+            # Display list of candidates.
+            print_(u'Finding tags for {0} "{1} - {2}".'.format(
+                u'track' if singleton else u'album',
+                item.artist if singleton else cur_artist,
+                item.title if singleton else cur_album,
+            ))
+
+            print_(u'Candidates:')
+            for i, match in enumerate(candidates):
+                # Index, metadata, and distance.
+                line = [
+                    u'{0}.'.format(i + 1),
+                    u'{0} - {1}'.format(
+                        match.info.artist,
+                        match.info.title if singleton else match.info.album,
+                    ),
+                    u'({0})'.format(dist_string(match.distance)),
+                ]
+
+                # Penalties.
+                penalties = penalty_string(match.distance, 3)
+                if penalties:
+                    line.append(penalties)
+
+                # Disambiguation
+                disambig = disambig_string(match.info)
+                if disambig:
+                    line.append(ui.colorize('text_highlight_minor',
+                                            '(%s)' % disambig))
+
+                print_(' '.join(line))
+
+            # Ask the user for a choice.
+            if singleton:
+                opts = ('Skip', 'Use as-is', 'Enter search', 'enter Id',
+                        'aBort')
+            else:
+                opts = ('Skip', 'Use as-is', 'as Tracks', 'Group albums',
+                        'Enter search', 'enter Id', 'aBort')
+            sel = ui.input_options(opts, numrange=(1, len(candidates)))
+            if sel == 's':
+                return importer.action.SKIP
+            elif sel == 'u':
+                return importer.action.ASIS
+            elif sel == 'm':
+                pass
+            elif sel == 'e':
+                return importer.action.MANUAL
+            elif sel == 't':
+                assert not singleton
+                return importer.action.TRACKS
+            elif sel == 'b':
+                raise importer.ImportAbort()
+            elif sel == 'i':
+                return importer.action.MANUAL_ID
+            elif sel == 'g':
+                return importer.action.ALBUMS
+            else:  # Numerical selection.
+                match = candidates[sel - 1]
+                if sel != 1:
+                    # When choosing anything but the first match,
+                    # disable the default action.
+                    require = True
+        bypass_candidates = False
+
+        # Show what we're about to do.
+        if singleton:
+            show_item_change(item, match)
+        else:
+            show_change(cur_artist, cur_album, match)
+
+        # Exact match => tag automatically if we're not in timid mode.
+        if rec == Recommendation.strong and not config['import']['timid']:
+            return match
+
+        # Ask for confirmation.
+        if singleton:
+            opts = ('Apply', 'More candidates', 'Skip', 'Use as-is',
+                    'Enter search', 'enter Id', 'aBort')
+        else:
+            opts = ('Apply', 'More candidates', 'Skip', 'Use as-is',
+                    'as Tracks', 'Group albums', 'Enter search', 'enter Id',
+                    'aBort')
+        default = config['import']['default_action'].as_choice({
+            'apply': 'a',
+            'skip': 's',
+            'asis': 'u',
+            'none': None,
+        })
+        if default is None:
+            require = True
+        sel = ui.input_options(opts, require=require, default=default)
+        if sel == 'a':
+            return match
+        elif sel == 'g':
+            return importer.action.ALBUMS
+        elif sel == 's':
+            return importer.action.SKIP
+        elif sel == 'u':
+            return importer.action.ASIS
+        elif sel == 't':
+            assert not singleton
+            return importer.action.TRACKS
+        elif sel == 'e':
+            return importer.action.MANUAL
+        elif sel == 'b':
+            raise importer.ImportAbort()
+        elif sel == 'i':
+            return importer.action.MANUAL_ID
+
+
+def manual_search(singleton):
+    """Input either an artist and album (for full albums) or artist and
+    track name (for singletons) for manual search.
+    """
+    artist = input_('Artist:')
+    name = input_('Track:' if singleton else 'Album:')
+    return artist.strip(), name.strip()
+
+
+def manual_id(singleton):
+    """Input an ID, either for an album ("release") or a track ("recording").
+    """
+    prompt = u'Enter {0} ID:'.format('recording' if singleton else 'release')
+    return input_(prompt).strip()
+
+
+class TerminalImportSession(importer.ImportSession):
+    """An import session that runs in a terminal.
+    """
+    def choose_match(self, task):
+        """Given an initial autotagging of items, go through an interactive
+        dance with the user to ask for a choice of metadata. Returns an
+        AlbumMatch object, ASIS, or SKIP.
+        """
+        # Show what we're tagging.
+        print_()
+        print_(displayable_path(task.paths, u'\n') +
+               u' ({0} items)'.format(len(task.items)))
+
+        # Take immediate action if appropriate.
+        action = _summary_judment(task.rec)
+        if action == importer.action.APPLY:
+            match = task.candidates[0]
+            show_change(task.cur_artist, task.cur_album, match)
+            return match
+        elif action is not None:
+            return action
+
+        # Loop until we have a choice.
+        candidates, rec = task.candidates, task.rec
+        while True:
+            # Ask for a choice from the user.
+            choice = choose_candidate(
+                candidates, False, rec, task.cur_artist, task.cur_album,
+                itemcount=len(task.items)
+            )
+
+            # Choose which tags to use.
+            if choice in (importer.action.SKIP, importer.action.ASIS,
+                          importer.action.TRACKS, importer.action.ALBUMS):
+                # Pass selection to main control flow.
+                return choice
+            elif choice is importer.action.MANUAL:
+                # Try again with manual search terms.
+                search_artist, search_album = manual_search(False)
+                _, _, candidates, rec = autotag.tag_album(
+                    task.items, search_artist, search_album
+                )
+            elif choice is importer.action.MANUAL_ID:
+                # Try a manually-entered ID.
+                search_id = manual_id(False)
+                if search_id:
+                    _, _, candidates, rec = autotag.tag_album(
+                        task.items, search_id=search_id
+                    )
+            else:
+                # We have a candidate! Finish tagging. Here, choice is an
+                # AlbumMatch object.
+                assert isinstance(choice, autotag.AlbumMatch)
+                return choice
+
+    def choose_item(self, task):
+        """Ask the user for a choice about tagging a single item. Returns
+        either an action constant or a TrackMatch object.
+        """
+        print_()
+        print_(task.item.path)
+        candidates, rec = task.candidates, task.rec
+
+        # Take immediate action if appropriate.
+        action = _summary_judment(task.rec)
+        if action == importer.action.APPLY:
+            match = candidates[0]
+            show_item_change(task.item, match)
+            return match
+        elif action is not None:
+            return action
+
+        while True:
+            # Ask for a choice.
+            choice = choose_candidate(candidates, True, rec, item=task.item)
+
+            if choice in (importer.action.SKIP, importer.action.ASIS):
+                return choice
+            elif choice == importer.action.TRACKS:
+                assert False  # TRACKS is only legal for albums.
+            elif choice == importer.action.MANUAL:
+                # Continue in the loop with a new set of candidates.
+                search_artist, search_title = manual_search(True)
+                candidates, rec = autotag.tag_item(task.item, search_artist,
+                                                   search_title)
+            elif choice == importer.action.MANUAL_ID:
+                # Ask for a track ID.
+                search_id = manual_id(True)
+                if search_id:
+                    candidates, rec = autotag.tag_item(task.item,
+                                                       search_id=search_id)
+            else:
+                # Chose a candidate.
+                assert isinstance(choice, autotag.TrackMatch)
+                return choice
+
+    def resolve_duplicate(self, task, found_duplicates):
+        """Decide what to do when a new album or item seems similar to one
+        that's already in the library.
+        """
+        log.warn(u"This {0} is already in the library!",
+                 ("album" if task.is_album else "item"))
+
+        # skip empty albums (coming from a previous failed import session)
+        if task.is_album:
+            real_duplicates = [dup for dup in found_duplicates if dup.items()]
+            if not real_duplicates:
+                log.info("All duplicates are empty, we ignore them")
+                task.should_remove_duplicates = True
+                return
+        else:
+            real_duplicates = found_duplicates
+
+        if config['import']['quiet']:
+            # In quiet mode, don't prompt -- just skip.
+            log.info(u'Skipping.')
+            sel = 's'
+        else:
+            # Print some detail about the existing and new items so the
+            # user can make an informed decision.
+            for duplicate in real_duplicates:
+                print_("Old: " + summarize_items(
+                    list(duplicate.items()) if task.is_album else [duplicate],
+                    not task.is_album,
+                ))
+
+            if real_duplicates != found_duplicates:  # there's empty albums
+                count = len(found_duplicates) - len(real_duplicates)
+                print_("Old: {0} empty album{1}".format(
+                       count, "s" if count > 1 else ""))
+
+            print_("New: " + summarize_items(
+                task.imported_items(),
+                not task.is_album,
+            ))
+
+            sel = ui.input_options(
+                ('Skip new', 'Keep both', 'Remove old')
+            )
+
+        if sel == 's':
+            # Skip new.
+            task.set_choice(importer.action.SKIP)
+        elif sel == 'k':
+            # Keep both. Do nothing; leave the choice intact.
+            pass
+        elif sel == 'r':
+            # Remove old.
+            task.should_remove_duplicates = True
+        else:
+            assert False
+
+    def should_resume(self, path):
+        return ui.input_yn(u"Import of the directory:\n{0}\n"
+                           "was interrupted. Resume (Y/n)?"
+                           .format(displayable_path(path)))
